requirements.txt
streamlit>=1.30
pytest>=7.0
rich>=13.0
google-genai>=1.0.0
python-dotenv>=1.0.0
